add organization ID to product

+++ Business/Handlers/Products/Commands/DeleteProductCommand.cs
@@ -9,7 +9,8 @@ using DataAccess.Abstract;
 using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
-
+using Core.Extensions;
+using Core.CrossCuttingConcerns.Context;
 
 namespace Business.Handlers.Products.Commands
 {
@@ -24,11 +25,13 @@ namespace Business.Handlers.Products.Commands
         {
             private readonly IProductRepository _productRepository;
             private readonly IMediator _mediator;
+            private readonly BabuAppContext _context;
 
-            public DeleteProductCommandHandler(IProductRepository productRepository, IMediator mediator)
+            public DeleteProductCommandHandler(IProductRepository productRepository, IMediator mediator, IAppContextService appContextService)
             {
                 _productRepository = productRepository;
                 _mediator = mediator;
+                _context = appContextService.GetAppContext();
             }
 
             [CacheRemoveAspect("Get")]
@@ -36,7 +39,12 @@ namespace Business.Handlers.Products.Commands
             [SecuredOperation(Priority = 1)]
             public async Task<IResult> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
             {
-                var productToDelete = _productRepository.Get(p => p.Id == request.Id);
+                var productToDelete = await _productRepository.GetAsync(p => p.Id == request.Id && p.OrganizationId == _context.OrganizationId);
+
+                if (productToDelete == null)
+                {
+                    return new ErrorResult(Messages.ProductNotFound);
+                }
 
                 _productRepository.Delete(productToDelete);
                 await _productRepository.SaveChangesAsync();
